fix: deprects CustomProvider, returns wrapped children from component map instead
